requirements.txt
discord.py==2.3.2
APScheduler==3.10.4
peewee==3.17.1
icalendar==5.0.11
python-dotenv==1.0.0
requests==2.32.3
